requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
pycodestyle==2.7.0
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
